feat(app-generic): refactor in memory providers provisions
