+++ packages/application-generic/src/services/in-memory-provider/index.ts
@@ -1,1 +1,2 @@
 export * from './in-memory-provider.service';
+export * from './types';
