+++ packages/application-generic/src/services/in-memory-provider/providers/redis-provider.ts
@@ -1,7 +1,7 @@
-import Redis, { RedisOptions } from 'ioredis';
+import Redis, { RedisOptions, ScanStream } from 'ioredis';
 import { ConnectionOptions } from 'tls';
 
-export { Redis, RedisOptions };
+export { Redis, RedisOptions, ScanStream };
 
 export const CLIENT_READY = 'ready';
 const DEFAULT_TTL_SECONDS = 60 * 60 * 2;
@@ -86,3 +86,6 @@ export const getRedisInstance = (): Redis | undefined => {
 
   return undefined;
 };
+
+export const isClientReady = (status: string): boolean =>
+  status === CLIENT_READY;
